4_function_move_data_to_dataset_big_query_with_transformatrions/requirements.txt
google-cloud-bigquery
google-cloud-kms
pandas-gbq
